🗃️ Add access rules & clean stuff

+++ pages/api/projects/deleteProject.ts
@@ -3,17 +3,17 @@ const Config = require('../../../api-lib/models/websiteconfig')
 
 import type { NextApiRequest, NextApiResponse } from 'next'
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-   await connectMongo()
-   let { name } = req.body
-   console.log('deleteProjct', name)
-   try {
-      let project = await Config.findOneAndDelete({ name: name }).clone()
+  await connectMongo()
+  let { name, email } = req.body
+  console.log('deleteProjct', name)
+  try {
+    let project = await Config.findOneAndDelete({ name: name, owner: email }).clone()
 
-      res.status(200).json('Success')
-   } catch (err) {
-      console.error(err.message)
-      res.status(500).json('Server error')
-   }
+    res.status(200).json('Success')
+  } catch (err) {
+    console.error(err.message)
+    res.status(500).json('Server error')
+  }
 }
 
 export default handler
